angle fixed.
next step : dist in double, for imprecisions?
compilation OK

+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: pkorsako <pkorsako@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/20 17:15:52 by pkorsako          #+#    #+#             */
-/*   Updated: 2024/03/13 15:54:27 by pkorsako         ###   ########.fr       */
+/*   Updated: 2024/03/15 17:56:37 by ajeannin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -177,8 +177,8 @@ int				get_value(unsigned int value, const char *flag);
 void			view_stocked_image(unsigned int *stock);
 void			view_stocked_col(unsigned int stock);
 void			max_dist(t_game *game);
-int				get_ipx(t_player *player, double i_py, double angle);
-int				get_ipy(t_player *player, double i_px, double angle);
+double				get_ipx(t_player *player, double i_py, double angle);
+double				get_ipy(t_player *player, double i_px, double angle);
 char			try_get_texture(char **map, int yt, int xt);
 
 //////////// Maths_tests ///////////////////
